DATA_GENERATOR/requirements.txt
# FloatChart Data Generator - Dependencies

# Data Processing
pandas
numpy

# Network & Data Fetching
requests
xarray
netCDF4

# Database
sqlalchemy
psycopg2-binary

# Environment
python-dotenv
